Quests: Controllerified
still a WIP, will use a Quest eloquent model to move this along.
Fixes #455 .
